% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte-carlo.R
\name{all.equal.montecarlo}
\alias{all.equal.montecarlo}
\title{todo: have args to check only results....
 should also check that dgp_to_poi() function body is identical?
this ignores some attributes.
e.g., should not compare system times
If 'target' has the same ptups, but in a different order, all.equal() will return false by default.
If you do not want to be sensitive to order, you can use argument "sensitive_to_pn_order".}
\usage{
\method{all}{equal.montecarlo}(
  target,
  current,
  tolerance = sqrt(.Machine$double.eps),
  sensitive_to_pn_order = TRUE,
  sensitive_to_stat_order = TRUE,
  verbose = 1,
  ...
)
}
\arguments{
\item{target}{The target MC to compare to.}

\item{current}{The current MC.}

\item{tolerance}{The amount by which a number in the current MC's stats_m can differ from the corresponding target MC's stats_m.}

\item{sensitive_to_pn_order}{A logical. If TRUE, return FALSE if order of current's pn-chunks are different from target's pn-chunks. Analogous to whether c(5, 11) is "equal" to c(11, 5).}

\item{sensitive_to_stat_order}{A logical. If TRUE, return FALSE if order of current's 'stats_m's are different from target's 'stats_m's.}

\item{verbose}{An integer (default of 1) to control the level feedback. Set to 0 to turn off messages that are not warnings or errors. At value 1, basic progress information is printed; at values 2 and higher, increasingly more messages are printed that are useful when debugging an unexpected result.}

\item{...}{Not used.}
}
\description{
todo: have args to check only results....
 should also check that dgp_to_poi() function body is identical?
this ignores some attributes.
e.g., should not compare system times
If 'target' has the same ptups, but in a different order, all.equal() will return false by default.
If you do not want to be sensitive to order, you can use argument "sensitive_to_pn_order".
}
