% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{construct_fX}
\alias{construct_fX}
\title{Construct f(X)}
\usage{
construct_fX(beta, formula, ...)
}
\arguments{
\item{beta}{The vector of coefficients.}

\item{formula}{The formula object.}

\item{...}{The objects used to construct the data (passed to construct_rhs_df).}
}
\description{
TODO: rename this to 'construct_f_of_x' to avoid thinking this makes the partial derivative of f with respect to x (i.e., f_x).
}
