% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{construct_rhs_df}
\alias{construct_rhs_df}
\title{Construct a data.frame from RHS of formula}
\usage{
construct_rhs_df(formula, ...)
}
\arguments{
\item{formula}{The formula object.}

\item{...}{The data arguments.}
}
\description{
the LHS variable of 'form' does not need to be available.
}
