% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diag_ht_pval2rr_gen}
\alias{diag_ht_pval2rr_gen}
\title{A generator of diagnostics of p-value statistics}
\usage{
diag_ht_pval2rr_gen(p)
}
\arguments{
\item{p}{The level corresponding to the hypothesis test.}
}
\description{
A generator of diagnostics of p-value statistics
}
