% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diag_ht_rr}
\alias{diag_ht_rr}
\title{Diagnostic for rejection statistics}
\usage{
diag_ht_rr(stats_m, true_poi)
}
\arguments{
\item{stats_m}{the matrix of statistic values.}

\item{true_poi}{the true POI.}
}
\description{
Diagnostic for rejection statistics
}
