% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diag_prop_NA}
\alias{diag_prop_NA}
\title{Diagnostic for proportion NA}
\usage{
diag_prop_NA(stats_m, true_poi)
}
\arguments{
\item{stats_m}{The stats_m of an MC pn-pair.}

\item{true_poi}{The parameter of interest.}
}
\description{
Diagnostic for proportion NA. Useful in conjunction with statistics wrapped in ret_na_if_error.
}
