% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diags_est}
\alias{diags_est}
\alias{diag_est_bias}
\alias{diag_est_l1}
\alias{diag_est_l2}
\alias{diag_est_sq_error}
\alias{diag_est_abs_error}
\title{Diagnostics for statistics that are (typically point) estimators.}
\usage{
diag_est_bias(est, true_poi)

diag_est_l1(est, true_poi)

diag_est_l2(est, true_poi)

diag_est_sq_error(est, true_poi)

diag_est_abs_error(est, true_poi)
}
\arguments{
\item{est}{The vector of point estimates.}

\item{true_poi}{The vector of parameter values.}
}
\description{
These diagnostics can be used to calculate the (empirical) MSE. They can be combined with CI diags (e.g., see gen_diags_ci()).
}
\seealso{
See diags_ci for info on CI diagnostics. See gen_diags_ci() to create diagnostic functions for estimators with confidence intervals. See est_diag_to_ci_diag() for how to use a estimator diag as a CI diag.
}
