% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{gen_diags_ci}
\alias{gen_diags_ci}
\title{Generates MC diagnostics for estimators with confidence intervals}
\usage{
gen_diags_ci(
  diags_ci_l = list(diag_ci_length, diag_ci_covers),
  diags_est_l = list(diag_est_bias, diag_est_sq_error)
)
}
\arguments{
\item{diags_ci_l}{A list of individual CI diagnostics. Each function in this list should accept arguments "lower", "est", "upper", and "true_poi". See ?diags_ci}

\item{diags_est_l}{A list of individual estimator diagnostics. Each function in this list should accept arguments "est" and "true_poi". See ?diags_est}
}
\description{
Generates MC diagnostics for estimators with confidence intervals
}
\details{
Returns diagnostics that can be used in the 'diagnostics' argument of mc_run() or mc_diags().
}
