% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{gen_hook_save_partial}
\alias{gen_hook_save_partial}
\title{hook for saving partial results}
\usage{
gen_hook_save_partial(
  filename = "mc-bu.Rds",
  rm_if_no_error = TRUE,
  snapshot_dir = NA
)
}
\arguments{
\item{filename}{The filename to use. You may choose to save the file to a directory with cloud storage (e.g., ownCloud, Dropbox) if you want robustness respect to your computer not booting after a crash.}

\item{rm_if_no_error}{If all simulations succeed, delete the partial MC file. If you leave this as TRUE, you may want to save the return of mc_run() as soon as it succeeds.}

\item{snapshot_dir}{An optional snapshot directory, i.e., the return of mc_archive_new(). This feature is useful if you want to save the partial MC to the snapshot directory rather than the current working directory.}
}
\description{
The partial results file is updated to save partial MC results after each pn-chunk finishes (i.e., all of the simulations for a particular combination of a dgp_param and n). See details for the specifics of the partial results. After all simulations succeed, the file is deleted so you may want to immediatedly save the return of mc_run(). If you are concerned about an error after the last simulation (e.g., an error from a hook that comes after this hook), you can set the argument 'rm_if_no_error' to 'FALSE'.
}
\details{
the partial result files are useful for several reasons:
- if your computer loses power during a long simulation
- if your simulation run is more computationally intense than you realize and because of RAM
  R exits.
- if you want to view partial result tables (e.g., a PDF table can be automatically generated).
- ...
The backup mechanism relying on partial results is robust to crashes during the write process. That is, the backup file saved after the previous simulation completes will only be overwritten if...
}
