% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rounding.R
\name{gen_rounder_comparison}
\alias{gen_rounder_comparison}
\title{A rounder that chooses digits based on comparing entries in various ways}
\usage{
gen_rounder_comparison(
  compare_across_nobs = TRUE,
  compare_across_stats = TRUE,
  compare_across_dgpps = TRUE
)
}
\arguments{
\item{compare_across_nobs}{Compare across number of observations.}

\item{compare_across_stats}{Compare across number of stats.}

\item{compare_across_dgpps}{Compare across DGPPs (not yet implemented).}
}
\description{
This is a work in progress. Do not use yet.
}
