% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rounding.R
\name{gen_rounder_fixed}
\alias{gen_rounder_fixed}
\title{Generate a fixed-digit rounder}
\usage{
gen_rounder_fixed(digits)
}
\arguments{
\item{digits}{The number of digits to round to.}
}
\description{
Generate a fixed-digit rounder
}
