% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc-utilities.R
\name{get_core_mc_args}
\alias{get_core_mc_args}
\title{Get core MC arguments.}
\usage{
get_core_mc_args(mc)
}
\arguments{
\item{mc}{An MC object (e.g., created from mc_run).}
}
\description{
Get core MC arguments, e.g., as pass to mc_run().
}
