% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc-utilities.R
\name{get_nsims}
\alias{get_nsims}
\title{Get number of simulations of a pn-pair}
\usage{
get_nsims(mcx, pn_pair)
}
\arguments{
\item{mcx}{A "montecarlo" object (e.g., created from mc_run), or an "mcdiags" object (e.g., created from mc_diags).}

\item{pn_pair}{The pn_pair of the MC or mcdiags.}
}
\description{
Get number of simulations of a pn-pair for an MC object or mcdiags object.
}
