% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc-utilities.R
\name{get_nsims_vec}
\alias{get_nsims_vec}
\title{Get number of simulations for an MC}
\usage{
get_nsims_vec(mcx)
}
\arguments{
\item{mcx}{A "montecarlo" object (e.g., created from mc_run), or an "mcdiags" object (e.g., created from mc_diags).}
}
\description{
Get number of simulations for each pn pair of an MC
}
