% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc-utilities.R
\name{get_pn_diags_result}
\alias{get_pn_diags_result}
\title{Get diagnostics result of a pn-pair}
\usage{
get_pn_diags_result(mc, pn_pair)
}
\arguments{
\item{mc}{An MC object (e.g., created from mc_run).}

\item{pn_pair}{The pn_pair of the MC.}
}
\description{
Get diagnostics result of a pn-pair
}
