% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc-utilities.R
\name{get_pn_pair_most_recent}
\alias{get_pn_pair_most_recent}
\title{Get the most recently run pn-pair of an MC}
\usage{
get_pn_pair_most_recent(mc)
}
\arguments{
\item{mc}{An MC object (e.g., created from mc_run).}
}
\description{
Get the most recently run pn-pair of an MC. This function is useful in MC hooks.
}
