% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc-utilities.R
\name{get_pn_pairs}
\alias{get_pn_pairs}
\title{Get list of pn pairs from a Monte Carlo object}
\usage{
get_pn_pairs(mc, sortby = "none")
}
\arguments{
\item{mc}{An MC object (e.g., created from mc_run).}

\item{sortby}{The method to sort by, one of "none", "phash", "nobs", "time_finished".}
}
\description{
Get list of pn pairs from a Monte Carlo object
}
