% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{hook_print_partial}
\alias{hook_print_partial}
\title{Hook to print partial results
If verbose >= 1 and user provides diagnostics, calculate diagnostics for pn-chunks as soon as they're available and print them.}
\usage{
hook_print_partial(
  mc_part_done,
  mc_args_next,
  user_args,
  pn_pair_next,
  nsims_dispatched
)
}
\arguments{
\item{mc_part_done}{An MC object containing the pn-pairs of finished simulations so far.}

\item{mc_args_next}{The arguments that will be used for the next pn-chunk.}

\item{user_args}{The arguments given to mc_run() by the user.}

\item{pn_pair_next}{The pn-pair whose simulations will be run next.}

\item{nsims_dispatched}{The number of simulations dispatched so far (not necessarily completed).}
}
\description{
Hook to print partial results
If verbose >= 1 and user provides diagnostics, calculate diagnostics for pn-chunks as soon as they're available and print them.
}
