% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{hook_set_seed_factory}
\alias{hook_set_seed_factory}
\title{Produce a hook to set seed and rng}
\usage{
hook_set_seed_factory(mc, rnginfo_after = FALSE)
}
\arguments{
\item{mc}{An MC object (e.g., created from mc_run).}

\item{rnginfo_after}{If TRUE, set the RNG state to the RNG state from *after* the MC object was run (useful for extending the MC), otherwise use the rnginfo from before the MC object was run (useful for reproducing).}
}
\description{
This hook factory is used by mc_reproduce
}
