% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte-carlo.R
\docType{data}
\name{hooks_pnchunk_default}
\alias{hooks_pnchunk_default}
\title{Default list of mc hooks}
\format{
An object of class \code{list} of length 3.
}
\usage{
hooks_pnchunk_default
}
\description{
Default list of mc hooks, useful for appending.
}
\keyword{datasets}
