% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc-utilities.R
\name{is.mcdiags}
\alias{is.mcdiags}
\title{Reports whether x is an "mcdiags" object}
\usage{
is.mcdiags(x)
}
\arguments{
\item{x}{An object to test}
}
\description{
Reports whether x is an "mcdiags" object
}
