% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc-utilities.R
\name{is.montecarlo}
\alias{is.montecarlo}
\title{Reports whether x is a "montecarlo" object}
\usage{
is.montecarlo(x)
}
\arguments{
\item{x}{An object to test}
}
\description{
Reports whether x is a "montecarlo" object
}
