% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc-utilities.R
\name{list_to_mc}
\alias{list_to_mc}
\title{This function can be used to convert from different formats to montetools format.}
\usage{
list_to_mc(l, all_pairs)
}
\arguments{
\item{l}{The list.}

\item{all_pairs}{All of the pairs to create.}
}
\description{
This function can be used to convert from different formats to montetools format.
}
