% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{mc_archive_ls}
\alias{mc_archive_ls}
\title{Begin a new MC archive entry}
\usage{
mc_archive_ls(archive_dir = getOption("montetools_archive_dir"), verbose = 1)
}
\arguments{
\item{archive_dir}{The directory where archive files are stored.}

\item{verbose}{An integer (default of 1) to control the level feedback. Set to 0 to turn off messages that are not warnings or errors. At value 1, basic progress information is printed; at values 2 and higher, increasingly more messages are printed that are useful when debugging an unexpected result.}
}
\description{
Begin a new MC archive entry
}
\details{
Returns the names of directories in the snapshot dir, sorted by mtime.
}
