% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{mc_archive_new}
\alias{mc_archive_new}
\title{Begin a new MC archive entry}
\usage{
mc_archive_new(
  name = NA,
  archive_dir = getOption("montetools_archive_dir"),
  verbose = 1
)
}
\arguments{
\item{name}{The name of the snapshot directory. The actual directory will suffix "_1" to it, or if that directory exists will keep incrementing the suffix.}

\item{archive_dir}{The directory where to store the archive files.}

\item{verbose}{An integer (default of 1) to control the level feedback. Set to 0 to turn off messages that are not warnings or errors. At value 1, basic progress information is printed; at values 2 and higher, increasingly more messages are printed that are useful when debugging an unexpected result.}
}
\description{
Begin a new MC archive entry
}
\details{
This function takes a snapshot of the repository state. The user should call this function at beginning of the script. That way, if the repository changes in the mean time (while the simulations are running), the snapshot taken at the beginning is still valid (i.e., an accurate representation of the repository at the time when the simulations were *started*).
}
