% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{mc_archive_store}
\alias{mc_archive_store}
\title{Store an mc in the mc_archive_dir.}
\usage{
mc_archive_store(mc, snapshot_dir)
}
\arguments{
\item{mc}{An MC object (e.g., created from mc_run).}

\item{snapshot_dir}{The snapshot directory, i.e., the return of mc_archive_new().}
}
\description{
Store an mc in the mc_archive_dir.
}
