% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.R
\name{mc_compare}
\alias{mc_compare}
\title{TODO: I think this function assumes that nsims of before are very high?
       e.g., do we reuire its values to be close to its plim?
To understand the default comparison mode, note that if before = after,
this function could still give warnings if 'sdtol' and nsims are both relatively
small because we look at whether the bounds of the *confidence interval* are
within sdtol of the mean of the other.}
\usage{
mc_compare(before, after, diagnostics, aggregators, sdtol = 1/5)
}
\arguments{
\item{before}{The "before" MC.}

\item{after}{The "after" MC.}

\item{diagnostics}{The diagnostics for comparison.}

\item{aggregators}{The aggregators (e.g., mean) of diagnostics results.}

\item{sdtol}{The tolerance, in units of SDs.}
}
\description{
TODO: I think this function assumes that nsims of before are very high?
       e.g., do we reuire its values to be close to its plim?
To understand the default comparison mode, note that if before = after,
this function could still give warnings if 'sdtol' and nsims are both relatively
small because we look at whether the bounds of the *confidence interval* are
within sdtol of the mean of the other.
}
