% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{mc_diagnostics}
\alias{mc_diagnostics}
\title{Diagnostic functions for evaluating Monte Carlo results}
\description{
Diagnostic functions for evaluating Monte Carlo results
}
\details{
An MC diagnostic, or 'diag' for short, is a function whose purpose
is to evaluate and summarize the results of MC simulations. A diag can be passed to the
'diagnostics' argument of `mc_run` to display results after each pn-chunk
finishes, or can be passed to `mc_diags()` to calculate diagnostics on a
completed MC object.

A diagnostic function must accept two arguments. The first argument, 'stats_m', is the matrix
of stacked (across simulations) statistics. The second argument, 'true_poi', is the true POI value, i.e., the value of the POI associated with the data generating process parameter (DGPP) that the DGP used to produce the simulated data set (often the true POI *is* the DGPP). You may also pass a list of such functions.

The diagnostic function should return a data.frame or matrix with column names the names of the diagnostics. No column name in the returned object should be empty.

Some common diagnostics are functions of only the simulated statistic values (e.g., average/median of rejection rates for MC hypothesis tests, or confidence band length for MC estimation), and some also input the dgp param (e.g., squared error for MC estimation).
}
\seealso{
See gen_diags_ci() to create diagnostic functions for estimators with confidence intervals.
}
