% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{mc_diags}
\alias{mc_diags}
\title{Run diagnostics on result of Monte Carlo simulations}
\usage{
mc_diags(mc, diagnostics = NULL, aggregators = NULL, verbose = 1)
}
\arguments{
\item{mc}{An MC object (e.g., created from mc_run).}

\item{diagnostics}{See '?mc_diagnostics' for information. Each diagnostics in the list will be run for each statistic for every pn-pair.}

\item{aggregators}{a list of functions to aggregate the simulation results. Typically this is left at the default of using the mean or changed to the median by specifying list(median = median).
By default (if aggregators is NULL), aggregators will report 'diag_prop_NA'.}

\item{verbose}{An integer (default of 1) to control the level feedback. Set to 0 to turn off messages that are not warnings or errors. At value 1, basic progress information is printed; at values 2 and higher, increasingly more messages are printed that are useful when debugging an unexpected result.}
}
\description{
Applies diagnostic and aggregates the results.
}
