% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte-carlo.R
\name{mc_extend}
\alias{mc_extend}
\title{Extend an existing MC object with more simulations}
\usage{
mc_extend(
  mc,
  nsims_additional = get_nsims_vec(mc),
  restore_globals = TRUE,
  verbose = 1,
  ...
)
}
\arguments{
\item{mc}{The MC object to create an extension from.}

\item{nsims_additional}{The number of simulations to extend by. This can be a vector. By default, we use the same number of simulations as the original 'mc' argument used (i.e., the number of simulations after extension will be double that of the original).}

\item{restore_globals}{logical for whether to restore the globals that were captured (if the argument 'store_globals' was TRUE) by mc_run().}

\item{verbose}{An integer (default of 1) to control the level feedback. Set to 0 to turn off messages that are not warnings or errors. At value 1, basic progress information is printed; at values 2 and higher, increasingly more messages are printed that are useful when debugging an unexpected result.}

\item{...}{Arguments that will be passed to the call to mc_run() (e.g., 'parallel = TRUE').}
}
\description{
mc_extend takes the result of a previous MC object created by mc_run(), and runs more simulations.
}
