% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc-utilities.R
\name{mc_first_n_sims}
\alias{mc_first_n_sims}
\title{Extract only the first 'n' sims of each pn-pair}
\usage{
mc_first_n_sims(mc, n)
}
\arguments{
\item{mc}{An MC object (e.g., created from mc_run).}

\item{n}{The number of simulations to limit the MC to.}
}
\description{
Useful to see if can reproduce just the first, e.g., 1, sim of each pn-pair.
}
