% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte-carlo.R
\name{mc_hooks}
\alias{mc_hooks}
\title{mc hooks}
\description{
mc hooks
}
\details{
mc hooks customize display of partial results and other things.

This type of hook takes arguments 'mc_part_done', 'mc_args_next', 'user_args', and 'pn_pair_next'.

See examples.
}
