% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulations.R
\name{mc_merge}
\alias{mc_merge}
\title{Combine/merge MC objects, matching on (n, phash) pairs.}
\usage{
mc_merge(..., verbose = 1)
}
\arguments{
\item{...}{The MC objects to be combined.}

\item{verbose}{An integer (default of 1) to control the level feedback. Set to 0 to turn off messages that are not warnings or errors. At value 1, basic progress information is printed; at values 2 and higher, increasingly more messages are printed that are useful when debugging an unexpected result.}
}
\description{
They can be any combination of MC objects or file names of .Rds files (each containing an MC object).
The first argument determines the "main" MC, whose attributes and logged system details will be used in the returned merged MC object; such elements of the other MC objects will not be preserved.
}
