% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc-utilities.R
\name{mc_stats_m}
\alias{mc_stats_m}
\title{Get the stats matrix of a 'montecarlo' object.}
\usage{
mc_stats_m(mc, pn_pair)
}
\arguments{
\item{mc}{An MC object (e.g., created from mc_run).}

\item{pn_pair}{The pn-pair of the MC object to get the stats matrix of.}
}
\description{
Get the stats matrix of a 'montecarlo' object.
}
