% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{montetools-package}
\alias{montetools-package}
\alias{montetools}
\title{A suite of tools for running and presenting Monte Carlo simulations}
\description{
A suite for facilitating tasks related to Monte Carlo simulations, including the running the simulations, extending them, reproducing them, and displaying results.
}
\author{
Scott Kostyshak <\url{https://people.clas.ufl.edu/skostyshak/}>
}
