% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ret_na_if_error}
\alias{ret_na_if_error}
\title{Return NA if error}
\usage{
ret_na_if_error(func, verbose = 1)
}
\arguments{
\item{func}{The function that will be wrapped.}

\item{verbose}{An integer (default of 1) to control the level feedback. Set to 0 to turn off messages that are not warnings or errors. At value 1, basic progress information is printed; at values 2 and higher, increasingly more messages are printed that are useful when debugging an unexpected result.}
}
\description{
Function wrapper to return NA if error.
}
