% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{try_or_na}
\alias{try_or_na}
\title{stop_on_these_classes: In some cases, we do usually want to give an error (and not
return NA). For example, it is usually nice to error out on
'packageNotFoundError'. That particular error occurs when running mc_reproduce
without a required package.}
\usage{
try_or_na(
  expr,
  na_on_these_messages = c("*"),
  stop_on_these_classes = c("packageNotFoundError"),
  stop_on_these_messages = c("could not find function"),
  verbose = 1
)
}
\arguments{
\item{expr}{The expression to try.}

\item{na_on_these_messages}{The pattern of messages to "allow", i.e., return NA without error. 'stop_on_these_messages' overrides these messages.}

\item{stop_on_these_classes}{The classes of error that should result in an error (i.e., not return NA).}

\item{stop_on_these_messages}{The pattern of messages that should cause an error (i.e., not return NA).}

\item{verbose}{An integer (default of 1) to control the level feedback. Set to 0 to turn off messages that are not warnings or errors. At value 1, basic progress information is printed; at values 2 and higher, increasingly more messages are printed that are useful when debugging an unexpected result.}
}
\description{
stop_on_these_classes: In some cases, we do usually want to give an error (and not
return NA). For example, it is usually nice to error out on
'packageNotFoundError'. That particular error occurs when running mc_reproduce
without a required package.
}
