% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte-carlo.R
\name{validate_montecarlo}
\alias{validate_montecarlo}
\title{maybe define the format in this help file?
export so other users/packages can build their own montecarlo objects
and take advantage of other functions in montetools (e.g., diags or reporting results).}
\usage{
validate_montecarlo(mc)
}
\arguments{
\item{mc}{An MC object (e.g., created from mc_run).}
}
\description{
maybe define the format in this help file?
export so other users/packages can build their own montecarlo objects
and take advantage of other functions in montetools (e.g., diags or reporting results).
}
