% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulations.R
\name{mc_stats_subset}
\alias{mc_stats_subset}
\title{Subset on stats}
\usage{
mc_stats_subset(mc, remove_these, keep_these)
}
\arguments{
\item{mc}{An MC object (e.g., created from mc_run).}

\item{remove_these}{The names of the statistics to remove. Cannot be combined with 'keep_these'.}

\item{keep_these}{The names of the statistics to keep. Cannot be combined with 'remove_these'.}
}
\description{
Note that the "sys_stats" simulation time is no longer valid after using
mc_stats_subset().
}
